+++ core/runner/src/main/java/com/cognifide/aet/runner/processing/steps/CollectionResultsRouter.java
@@ -25,7 +25,7 @@ import com.cognifide.aet.communication.api.metadata.Test;
 import com.cognifide.aet.communication.api.metadata.Url;
 import com.cognifide.aet.communication.api.queues.JmsConnection;
 import com.cognifide.aet.communication.api.util.ExecutionTimer;
-import com.cognifide.aet.runner.MessagesManager;
+import com.cognifide.aet.communication.api.queues.QueuesConstant;
 import com.cognifide.aet.runner.RunnerConfiguration;
 import com.cognifide.aet.runner.processing.TimeoutWatch;
 import com.cognifide.aet.runner.processing.data.wrappers.RunIndexWrapper;
@@ -184,12 +184,12 @@ public class CollectionResultsRouter extends StepManager implements TaskFinishPo
 
   @Override
   protected String getQueueInName() {
-    return MessagesManager.createFullQueueName("collectorResults");
+    return QueuesConstant.COLLECTOR.getResultsQueueName();
   }
 
   @Override
   protected String getQueueOutName() {
-    return MessagesManager.createFullQueueName("comparatorJobs");
+    return QueuesConstant.COMPARATOR.getJobsQueueName();
   }
 
   @Override